Add checksum recalculation in background.
Fix bug in checksum calculation.
New release: 2.5.2.

+++ src/metarestore/merger.h
@@ -7,4 +7,4 @@
 #include <vector>
 
 int merger_start(const std::vector<std::string>& filenames, uint64_t maxhole);
-int merger_loop(void);
+uint8_t merger_loop(void);
